// src/features/mapa/types/routing-type.ts
export interface RoutingItem {
  company: string;
  route: string;
  transport: string;
  licensePlate: string;
  shipment: string;
  sequence: number;
  customer: string;
  customerName: string;
  location: string;
  routedVehicle: string;
  effectiveVehicle: string;
  carrier: string;
  cargoType: string;
} 